requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.34.0
supabase==2.7.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.8.0
python-jose==3.3.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.143.0
pytz==2024.1
